refactor(data): 优化数据处理和模型定义
- 引入 safe_date 工具函数统一处理日期格式转换
- 修改财务数据脚本中的日期处理逻辑
- 为多个数据库表添加唯一约束以确保数据完整性
- 更新回测相关模型和架构定义
- 调整因子模型的字段定义和数据类型
- 优化 PIT 守护逻辑中的日期比较方式
- 修改市场数据模型的字段精度和索引配置
- 重构模型相关架构定义和字段类型
- 更新因子计算服务的实际计算逻辑
- 调整指数和行业数据模型的数据类型精度

+++ app/models/market/index_basic.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Date, DateTime, Float, Integer, String
+from sqlalchemy import Column, Date, DateTime, Integer, Numeric, String
 from sqlalchemy.sql import func
 
 from app.db.base_class import Base
@@ -16,7 +16,7 @@ class IndexBasic(Base):
     publisher = Column(String(50))
     category = Column(String(20))
     base_date = Column(Date)
-    base_point = Column(Float)
+    base_point = Column(Numeric(20, 4))
     list_date = Column(Date)
     created_at = Column(DateTime, server_default=func.now())
     updated_at = Column(DateTime, server_default=func.now(), onupdate=func.now())
