+++ app/models/market/index_components.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Date, DateTime, Float, Index, Integer, String
+from sqlalchemy import Column, Date, DateTime, Index, Integer, Numeric, String
 from sqlalchemy.sql import func
 
 from app.db.base_class import Base
@@ -17,5 +17,5 @@ class IndexComponent(Base):
     index_code = Column(String(20), nullable=False)
     trade_date = Column(Date, nullable=False)
     ts_code = Column(String(20), nullable=False)
-    weight = Column(Float)
+    weight = Column(Numeric(10, 6))
     created_at = Column(DateTime, server_default=func.now())
